requirements.txt
pandas==1.3.5
requests==2.28.1
seaborn==0.12.1
fpdf==1.7.2
bs4==0.0.1
apache-airflow==2.5.0